Replaced interface enum with abstract class and implementation
